Notes Domain Refactor (#11)
* Arrange by note domain

* Update package and imports

* Update NoteActionHandler.java

* Update NoteDao.java

* update tests

* remove redundant imports? (since same package)
