+++ src/main/java/client/processor/npc/FredrickProcessor.java
@@ -32,11 +32,11 @@ import client.inventory.ItemFactory;
 import client.inventory.manipulator.InventoryManipulator;
 import net.server.Server;
 import net.server.world.World;
+import note.NoteService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import server.ItemInformationProvider;
 import server.maps.HiredMerchant;
-import service.NoteService;
 import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
