+++ src/main/java/net/server/channel/handlers/PlayerLoggedinHandler.java
@@ -54,11 +54,11 @@ import net.server.guild.GuildPackets;
 import net.server.world.PartyCharacter;
 import net.server.world.PartyOperation;
 import net.server.world.World;
+import note.NoteService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scripting.event.EventInstanceManager;
 import server.life.MobSkill;
-import service.NoteService;
 import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
