+++ src/main/java/net/server/channel/handlers/RingActionHandler.java
@@ -34,11 +34,11 @@ import net.AbstractPacketHandler;
 import net.packet.InPacket;
 import net.server.channel.Channel;
 import net.server.world.World;
+import note.NoteService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scripting.event.EventInstanceManager;
 import server.ItemInformationProvider;
-import service.NoteService;
 import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
