+++ src/main/java/note/NoteRowMapper.java
@@ -1,6 +1,6 @@
-package database.note;
+package note;
 
-import model.Note;
+import note.Note;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
