+++ src/main/java/tools/DatabaseConnection.java
@@ -3,11 +3,11 @@ package tools;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import config.YamlConfig;
-import database.note.NoteRowMapper;
 import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.Jdbi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import note.NoteRowMapper;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
